+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Add `save_product_media`, `delete_product_media`, `save_configurable_product_option`, `save_base_prices`,
+  `get_special_prices`, `save_special_prices`, `delete_special_prices`, `delete_special_prices_by_sku`,
+  `create_category`, `ship_order`
 * Minor type hints improvements
 
 ## 1.0.0 (2021/12/21)